+++ app-includes/cota-search.php
@@ -6,7 +6,7 @@
 </head>
 <body>
     <h2>Search Family</h2>
-    <form action="../app-includes/edit_family.php" method="get">
+    <form action="../app-includes/cota-edit-family.php" method="get">
         <label>Enter Family Name:</label>
         <input type="text" name="familyname" required>
         <button type="submit">Search</button>